Start splitting Static API tests out
